requirements.txt
pydantic==1.10.7
pyyaml==6.0.1